chore(docs): Initial version of Docs
Most of this is straight from the RFC, but I added a couple lines and
images for explanation.

The Introduction Sections both will need some of the teams input to
introduce the "vision"

Also I'm using Prettier so it autoformatted the config file, but there
were no changes. I wasn't sure whether to include in the PR, but if the
team uses Prettier as well, this will make future PRs easier